feat(asset-selector): extract search logic to hook
